requirements.txt
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-aws>=0.2.0
boto3>=1.34.0
python-dotenv>=1.0.1
fastmcp>=3.0.0
